LayoutTests/imported/w3c/web-platform-tests/dom/events/non-cancelable-when-passive/passive-wheel-event-listener-on-div-expected.txt


FAIL passive wheel event listener on div assert_equals: expected false but got true
